+++ fl_amap/android/src/main/kotlin/fl/amap/AMapLocation.kt
@@ -5,12 +5,15 @@ import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.content.Context
 import android.graphics.Color
+import android.graphics.ColorSpace
 import android.os.Build
 import com.amap.api.location.AMapLocation
 import com.amap.api.location.AMapLocationClient
 import com.amap.api.location.AMapLocationClientOption
 import com.amap.api.location.AMapLocationListener
 import com.amap.api.location.AMapLocationQualityReport
+import com.amap.api.location.CoordinateConverter
+import com.amap.api.location.DPoint
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
@@ -76,7 +79,7 @@ class AMapLocation(plugin: FlutterPlugin.FlutterPluginBinding) : MethodChannel.M
 
             "getLocation" -> {
                 if (isLocation) {
-                    result.success(false)
+                    result.success(null)
                     return
                 }
                 try {
@@ -128,8 +131,53 @@ class AMapLocation(plugin: FlutterPlugin.FlutterPluginBinding) : MethodChannel.M
                 result.success(true)
             }
 
+            "isAMapDataAvailable" -> {
+                result.success(isAMapDataAvailable(call.arguments as Map<*, *>))
+            }
+
+            "calculateLineDistance" -> {
+                result.success(calculateLineDistance(call.arguments as Map<*, *>))
+            }
+
+            "coordinateConverter" -> {
+                result.success(coordinateConverter(call.arguments as Map<*, *>))
+            }
+
             else -> result.notImplemented()
         }
+    }
+
+    private fun isAMapDataAvailable(args: Map<*, *>): Boolean {
+        return CoordinateConverter.isAMapDataAvailable(
+            args["latitude"] as Double,
+            args["longitude"] as Double,
+        )
+    }
+
+    private fun calculateLineDistance(args: Map<*, *>): Float {
+        val start = DPoint(args["startLatitude"] as Double, args["startLongitude"] as Double)
+        val end = DPoint(args["endLatitude"] as Double, args["endLongitude1"] as Double)
+        return CoordinateConverter.calculateLineDistance(start, end)
+    }
+
+    private fun coordinateConverter(args: Map<*, *>): Map<String, Any?> {
+        val dPoint = DPoint()
+        dPoint.latitude = args["latitude"] as Double
+        dPoint.longitude = args["longitude"] as Double
+        val coordinateConverter = CoordinateConverter(context)
+        coordinateConverter.from(CoordinateConverter.CoordType.entries[args["from"] as Int])
+        try {
+            coordinateConverter.coord(dPoint)
+            val point = coordinateConverter.convert()
+            return mapOf(
+                "code" to 0, "latitude" to point.latitude, "longitude" to point.longitude
+            )
+        } catch (e: Exception) {
+            return mapOf(
+                "code" to 1, "message" to e.message
+            )
+        }
+
     }
 
     private fun buildNotification(args: Map<*, *>): Notification {
@@ -144,7 +192,17 @@ class AMapLocation(plugin: FlutterPlugin.FlutterPluginBinding) : MethodChannel.M
             channel.description = args["description"] as String?
             channel.lockscreenVisibility = args["lockscreenVisibility"] as Int
             channel.enableLights(args["enableLights"] as Boolean) //是否在桌面icon右上角展示小圆点
-            channel.lightColor = Color.parseColor(args["lightColor"] as String) //小圆点颜色
+            val lightColor = args["lightColor"] as Map<*, *>?
+            if (lightColor != null) {
+                val color = Color.valueOf(
+                    (lightColor["r"] as Double).toFloat(),
+                    (lightColor["g"] as Double).toFloat(),
+                    (lightColor["b"] as Double).toFloat(),
+                    (lightColor["a"] as Double).toFloat(),
+                    ColorSpace.get(ColorSpace.Named.entries[lightColor["colorSpace"] as Int]),
+                )
+                channel.lightColor = color.toArgb() //小圆点颜色
+            }
             channel.setShowBadge(args["showBadge"] as Boolean) //是否在久按桌面图标时显示此渠道的通知
             notificationManager.createNotificationChannel(channel)
             builder = Notification.Builder(context, channelId)
@@ -162,15 +220,15 @@ class AMapLocation(plugin: FlutterPlugin.FlutterPluginBinding) : MethodChannel.M
     private fun setLocationOption(arguments: Map<*, *>) {
         println(arguments)
         option.locationMode =
-            AMapLocationClientOption.AMapLocationMode.values()[arguments["locationMode"] as Int]
+            AMapLocationClientOption.AMapLocationMode.entries[arguments["locationMode"] as Int]
         val protocol =
-            AMapLocationClientOption.AMapLocationProtocol.values()[arguments["locationProtocol"] as Int]
+            AMapLocationClientOption.AMapLocationProtocol.entries[arguments["locationProtocol"] as Int]
         AMapLocationClientOption.setLocationProtocol(protocol)
         val locationPurpose = arguments["locationPurpose"] as Int?
         option.locationPurpose =
-            if (locationPurpose == null) null else AMapLocationClientOption.AMapLocationPurpose.values()[locationPurpose]
+            if (locationPurpose == null) null else AMapLocationClientOption.AMapLocationPurpose.entries[locationPurpose]
         option.geoLanguage =
-            AMapLocationClientOption.GeoLanguage.values()[(arguments["geoLanguage"] as Int)]
+            com.amap.api.location.AMapLocationClientOption.GeoLanguage.entries[(arguments["geoLanguage"] as Int)]
         option.isGpsFirst = arguments["gpsFirst"] as Boolean
         option.gpsFirstTimeout = (arguments["gpsFirstTimeout"] as Int).toLong()
         option.isMockEnable = arguments["mockEnable"] as Boolean