+++ fl_amap/lib/fl_amap.dart
@@ -1,3 +1,5 @@
+import 'dart:ui';
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -54,7 +56,10 @@ class LatLng {
       : latitude = map['latitude'] as double?,
         longitude = map['longitude'] as double?;
 
+  /// 维度
   double? latitude;
+
+  /// 经度
   double? longitude;
 
   Map<String, dynamic> toMap() =>