+++ fl_amap_map/example/lib/main.dart
@@ -2,9 +2,9 @@ import 'package:example/geo_fence_page.dart';
 import 'package:example/loaction_page.dart';
 import 'package:example/map_view_page.dart';
 import 'package:fl_amap_map/fl_amap_map.dart';
+import 'package:fl_extended/fl_extended.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_waya/flutter_waya.dart';
 import 'package:permission_handler/permission_handler.dart';
 
 Future<void> main() async {
@@ -19,8 +19,8 @@ Future<void> main() async {
     debugPrint('高德地图ApiKey设置$value');
   });
   runApp(MaterialApp(
-      navigatorKey: GlobalWayUI().navigatorKey,
-      scaffoldMessengerKey: GlobalWayUI().scaffoldMessengerKey,
+      navigatorKey: FlExtended().navigatorKey,
+      scaffoldMessengerKey: FlExtended().scaffoldMessengerKey,
       debugShowCheckedModeBanner: false,
       theme: ThemeData.light(useMaterial3: true),
       darkTheme: ThemeData.dark(useMaterial3: true),