+++ format.sh
@@ -1,0 +1,4 @@
+dart format fl_amap/lib
+dart format fl_amap/example/lib
+dart format fl_amap_map/lib
+dart format fl_amap_map/example/lib